TEMP: vendor AnsiCodeProcessor
